+++ cuprum/_subprocess_execution.py
@@ -15,7 +15,7 @@ import sys
 import time
 import typing as typ
 
-from cuprum._pipeline_internals import _EventDetails, _StageObservation
+from cuprum._pipeline_types import _EventDetails, _StageObservation
 from cuprum._process_lifecycle import _merge_env, _terminate_process
 from cuprum._streams import _consume_stream, _StreamConfig
 from cuprum._subprocess_context import _cwd_arg, _resolve_timeout, _sh_module